Booking Menu selesai, bagian cetak tiket belum

+++ src/server/controller/admin_panel/DeleteBooking.java
@@ -3,6 +3,7 @@ package server.controller.admin_panel;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import server.sessions.UserSession;
 
 public class DeleteBooking {
 
@@ -16,6 +17,11 @@ public class DeleteBooking {
     // DELETE BOOKING
     public void deleteByBookingCode(Connection conn, String bookingCode)
             throws SQLException, Exception {
+
+        String adminId = UserSession.getAdminId();
+        if (adminId == null) {
+            throw new Exception("Session admin tidak ditemukan!");
+        }
 
         String sql = "DELETE FROM booking WHERE booking_code = ?";
